+++ src/App.php
@@ -7,7 +7,7 @@ use Yaoi\Command\Definition;
 
 class App extends Command\Application
 {
-    public static $ver = 'v1.7.13';
+    public static $ver = 'v1.8.0';
 
     public $diff;
     public $apply;