+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.0] - 2020-09-30
+
+### Added
+- Dependencies updated.
+- Added schema patches to `gen-go` and `gen-php`.
+- Added property name fields controls in `gen-go`.
+
 ## [1.7.13] - 2020-09-26
 
 ### Added
@@ -138,6 +145,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Local file resolver in references.
 
+[1.8.0]: https://github.com/swaggest/json-cli/compare/v1.7.13...v1.8.0
 [1.7.13]: https://github.com/swaggest/json-cli/compare/v1.7.12...v1.7.13
 [1.7.12]: https://github.com/swaggest/json-cli/compare/v1.7.11...v1.7.12
 [1.7.11]: https://github.com/swaggest/json-cli/compare/v1.7.10...v1.7.11